lazy load modeling participation, extract markdown utils
