+++ src/main/webapp/app/shared/markdown-editor/commands/katex.command.ts
@@ -1,5 +1,5 @@
 import { DomainTagCommand } from 'app/shared/markdown-editor/domainCommands/domainTag.command';
-import { ArtemisMarkdown } from 'app/shared/markdown.service';
+import { addTextAtCursor } from 'app/shared/util/markdown-util';
 
 /**
  * Insert a katex compatible formula.
@@ -10,7 +10,7 @@ export class KatexCommand extends DomainTagCommand {
     buttonTranslationString = 'artemisApp.markdownEditor.commands.katex';
     execute(input?: string): void {
         const text = `${this.getOpeningIdentifier()}e^{\\frac{1}{4} y^2}${this.getClosingIdentifier()}`;
-        ArtemisMarkdown.addTextAtCursor(text, this.aceEditorContainer);
+        addTextAtCursor(text, this.aceEditorContainer);
     }
 
     getOpeningIdentifier(): string {
