+++ src/main/webapp/app/shared/markdown-editor/domainCommands/credits.command.ts
@@ -1,5 +1,5 @@
 import { DomainTagCommand } from 'app/shared/markdown-editor/domainCommands/domainTag.command';
-import { ArtemisMarkdown } from 'app/shared/markdown.service';
+import { addTextAtCursor } from 'app/shared/util/markdown-util';
 
 export class CreditsCommand extends DomainTagCommand {
     public static readonly identifier = '[credits]';
@@ -15,7 +15,7 @@ export class CreditsCommand extends DomainTagCommand {
      */
     execute(): void {
         const text = '\n' + this.getOpeningIdentifier() + CreditsCommand.text;
-        ArtemisMarkdown.addTextAtCursor(text, this.aceEditorContainer);
+        addTextAtCursor(text, this.aceEditorContainer);
     }
 
     /**
