+++ src/main/webapp/app/shared/markdown-editor/domainCommands/feedback.command.ts
@@ -1,5 +1,5 @@
 import { DomainTagCommand } from 'app/shared/markdown-editor/domainCommands/domainTag.command';
-import { ArtemisMarkdown } from 'app/shared/markdown.service';
+import { addTextAtCursor } from 'app/shared/util/markdown-util';
 
 export class FeedbackCommand extends DomainTagCommand {
     public static readonly identifier = '[feedback]';
@@ -14,7 +14,7 @@ export class FeedbackCommand extends DomainTagCommand {
      */
     execute(): void {
         const text = '\n' + this.getOpeningIdentifier() + FeedbackCommand.text;
-        ArtemisMarkdown.addTextAtCursor(text, this.aceEditorContainer);
+        addTextAtCursor(text, this.aceEditorContainer);
     }
 
     /**
