+++ src/main/webapp/app/shared/markdown-editor/domainCommands/gradingScaleCommand.ts
@@ -1,5 +1,5 @@
 import { DomainTagCommand } from 'app/shared/markdown-editor/domainCommands/domainTag.command';
-import { ArtemisMarkdown } from 'app/shared/markdown.service';
+import { addTextAtCursor } from 'app/shared/util/markdown-util';
 
 export class GradingScaleCommand extends DomainTagCommand {
     public static readonly identifier = '[gradingScale]';
@@ -14,7 +14,7 @@ export class GradingScaleCommand extends DomainTagCommand {
      */
     execute(): void {
         const text = '\n' + this.getOpeningIdentifier() + GradingScaleCommand.text;
-        ArtemisMarkdown.addTextAtCursor(text, this.aceEditorContainer);
+        addTextAtCursor(text, this.aceEditorContainer);
     }
 
     /**
