+++ src/main/webapp/app/shared/markdown-editor/domainCommands/instructionDescription.command.ts
@@ -1,5 +1,5 @@
 import { DomainTagCommand } from 'app/shared/markdown-editor/domainCommands/domainTag.command';
-import { ArtemisMarkdown } from 'app/shared/markdown.service';
+import { addTextAtCursor } from 'app/shared/util/markdown-util';
 
 export class InstructionDescriptionCommand extends DomainTagCommand {
     public static readonly identifier = '[description]';
@@ -14,7 +14,7 @@ export class InstructionDescriptionCommand extends DomainTagCommand {
      */
     execute(): void {
         const text = '\n' + this.getOpeningIdentifier() + InstructionDescriptionCommand.text;
-        ArtemisMarkdown.addTextAtCursor(text, this.aceEditorContainer);
+        addTextAtCursor(text, this.aceEditorContainer);
     }
 
     /**
