+++ src/main/webapp/app/shared/markdown.service.ts
@@ -2,51 +2,15 @@ import { Injectable } from '@angular/core';
 import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
 import * as showdown from 'showdown';
 import * as showdownKatex from 'showdown-katex';
-import * as ace from 'brace';
 import * as DOMPurify from 'dompurify';
-import { AceEditorComponent } from 'ng2-ace-editor';
 import { escapeStringForUseInRegex } from 'app/shared/util/global.utils';
 import { ExplanationCommand } from 'app/shared/markdown-editor/domainCommands/explanation.command';
 import { HintCommand } from 'app/shared/markdown-editor/domainCommands/hint.command';
 import { TextHintExplanationInterface } from 'app/entities/quiz/quiz-question.model';
 
-const Range = ace.acequire('ace/range').Range;
-
 @Injectable({ providedIn: 'root' })
 export class ArtemisMarkdown {
     static hintOrExpRegex = new RegExp(escapeStringForUseInRegex(`${ExplanationCommand.identifier}`) + '|' + escapeStringForUseInRegex(`${HintCommand.identifier}`), 'g');
-
-    /**
-     * adds the passed text into the editor of the passed ace editor component at the current curser by focusing, clearing a selection,
-     * moving the cursor to the end of the line, and finally inserting the given text.
-     * After that the new test will be selected
-     *
-     * @param text the text that will be added into the editor of the passed ace editor component
-     * @param aceEditorContainer holds the editor in which the text will be added at the current curser position
-     */
-    // TODO: we should move this functionality into another service or util class to break some dependencies on root level
-    static addTextAtCursor(text: String, aceEditorContainer: AceEditorComponent) {
-        aceEditorContainer.getEditor().focus();
-        aceEditorContainer.getEditor().clearSelection();
-        aceEditorContainer.getEditor().moveCursorTo(aceEditorContainer.getEditor().getCursorPosition().row, Number.POSITIVE_INFINITY);
-        aceEditorContainer.getEditor().insert(text);
-        const range = aceEditorContainer.getEditor().selection.getRange();
-        const commandIdentifier = text.split(']');
-        const offsetRange = commandIdentifier[0].length + 1;
-        range.setStart(range.start.row, offsetRange);
-        aceEditorContainer.getEditor().selection.setRange(range);
-    }
-
-    /**
-     * Remove the text at the specified range.
-     * @param from = col & row from which to start
-     * @param to = col & row at which to end
-     * @param aceEditorContainer
-     */
-    static removeTextRange(from: { col: number; row: number }, to: { col: number; row: number }, aceEditorContainer: AceEditorComponent) {
-        aceEditorContainer.getEditor().focus();
-        aceEditorContainer.getEditor().getSession().remove(new Range(from.row, from.col, to.row, to.col));
-    }
 
     constructor(private sanitizer: DomSanitizer) {}
